gui/requirements.txt
pyserial>=3.5
customtkinter>=5.2.0
pytest>=7.0
